D1 finished

+++ database/migrations/2023_07_06_093101_create_tweets_table.php
@@ -1,32 +1,55 @@
 <?php
 
+
+
+
 use Illuminate\Database\Migrations\Migration;
+
 use Illuminate\Database\Schema\Blueprint;
+
 use Illuminate\Support\Facades\Schema;
 
+
+
+
 return new class extends Migration
+
 {
+
     /**
+
      * Run the migrations.
+
      */
-    public function up()
-    {
-        Schema::table('tweets', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('text', 160);
-            $table->unsignedInteger('likes')->default(0);
-            $table->timestamps();
-        });
-    }
+
+     public function up()
+     {
+         if (!Schema::hasTable('tweets')) {
+             Schema::create('tweets', function (Blueprint $table) {
+                 $table->id();
+                 $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                 $table->string('text', 160);
+                 $table->unsignedInteger('likes')->default(0);
+                 $table->timestamps();
+             });
+         }
+     }
+     
 
 
 
     /**
+
      * Reverse the migrations.
+
      */
+
     public function down(): void
+
     {
+
         Schema::dropIfExists('tweets');
+
     }
+
 };
